# server/contracts/openapi.yml
...
openapi: 3.0.3
info:
  title: Notebook Wars metaserver REST API
  version: 0.1.0
  description: |
    Single source of truth for the REST contract (design-first, M15). The metaserver uses fastify-openapi-glue to wire routes + validation by operationId; the client uses openapi-typescript + openapi-fetch to generate typed fetch. Unified response envelope ApiResp<T>. See SERVER_API.md.
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
...
    AnalyticsEvent:
      type: object
      required:
        - event
        - ts
      properties:
        event:
          type: string
        ts:
          type: number
          description: Client unix ms
        props:
          type: object
          additionalProperties: true
    AnalyticsEventBatch:
      type: object
      required:
        - events
      properties:
        session_id:
          type: string
        device_id:
          type: string
        platform:
          type: string
          enum:
            - web
            - wechat
            - crazygames
        os:
          type: string
        game_version:
          type: string
        locale:
          type: string
        ua:
          type: string
        screen_w:
          type: integer
        screen_h:
          type: integer
        dpr:
          type: number
        consent:
          type: boolean
        events:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/AnalyticsEvent'
  responses:
    ErrorResp:
      description: Error envelope
      content:
        application/json:
          schema:
            type: object
            required:
              - ok
              - error
            properties:
              ok:
                type: boolean
                enum:
                  - false
              error:
                $ref: '#/components/schemas/Error'
paths:
...
  /bootstrap:
    get:
      operationId: bootstrap
      tags:
        - config
      summary: Public bootstrap config (FEATURE_FLAGS_DESIGN §9.3). Anonymous-callable (providing a token injects accountId for more precise evaluation); platform / publicId are passed via query. The server evaluates every allowlist entry and **returns only flags that differ from their default** (most players receive an empty map). Rules / allowlists are never sent — only boolean results. Targeted client log collection (client_log_*) uses this channel to deliver the matched log level.
      parameters:
        - name: platform
          in: query
          required: false
          schema:
            type: string
            enum:
              - web
              - wechat
              - crazygames
        - name: publicId
          in: query
          required: false
          description: Player-visible 9-digit publicId (for targeted evaluation); omitted = anonymous evaluation
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - data
                properties:
                  ok:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: object
                    required:
                      - flags
                    properties:
                      flags:
                        type: object
                        additionalProperties:
                          type: boolean
                        description: Contains only flags whose evaluated result differs from the default (key→bool); empty object for most players
                      paddleClientToken:
                        type: string
                        description: Paddle.js client-side token (COMMERCIAL_DESIGN §IAP client) — public, client-safe (ptok_/live_/test_). Only present when the server has NW_PADDLE_CLIENT_TOKEN configured; absent otherwise. The web client needs this to open the Paddle checkout overlay. Must be declared here or the Fastify response serializer strips it from the wire (fast-json-stringify only emits schema-declared properties).
  /client/log:
    post:
      operationId: clientLog
      tags:
        - config
      summary: 'Targeted client log forwarding → Loki (FEATURE_FLAGS_DESIGN §9.4). Only called by publicIds targeted by client_log_* (non-targeted clients receive an empty map on bootstrap and never report). The server re-validates targeting; unmatched logs are silently discarded. **Always returns 200** — Loki unreachable or not targeted still has no player impact. Labels: {source=client, level} only; publicId/tag/msg go inline (logfmt).'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - publicId
                - logs
              properties:
                publicId:
                  type: string
                  description: Reporter's 9-digit publicId (inline in Loki, for Grafana `| logfmt | publicId="..."` queries)
                platform:
                  type: string
                  enum:
                    - web
                    - wechat
                    - crazygames
                logs:
                  type: array
                  description: A batch of log entries (from the ring buffer at or above the match threshold); server takes at most the first 1000
                  items:
                    type: object
                    required:
                      - level
                      - msg
                      - ts
                    properties:
                      level:
                        type: string
                        enum:
                          - error
                          - warn
                          - info
                          - debug
                      msg:
                        type: string
                      ts:
                        type: number
                        description: epoch ms (client clock)
                      tag:
                        type: string
      responses:
        '200':
          description: Accepted (accepted = actual forwarded count; 0 if not targeted or invalid)
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - data
                properties:
                  ok:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: object
                    required:
                      - accepted
                    properties:
                      accepted:
                        type: integer
        '400':
          $ref: '#/components/responses/ErrorResp'
  /client/anomaly:
    post:
      operationId: clientAnomaly
      tags:
        - config
      summary: 'Bulk client anomaly event reporting → Loki (complements the targeted collection of /client/log, **not subject to allowPublicIds constraints**). Any client''s memory overrun / sustained CPU saturation (cpu) / WebGL context loss (webgl_lost) / main-loop hang (anr) / uncaught exception (jserror) / previous-session crash (crash) is reported directly, enabling in-the-wild anomaly diagnosis across all clients. Abuse protection: rate-limited per IP at 30 requests/60s (excess silently dropped), max first 200 entries taken, fields truncated. **Always returns 200** — Loki unreachable or rate-limited still has no player impact. Labels: {source=client, kind=anomaly} only; type/publicId/platform/detail/msg go inline (logfmt).'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - events
              properties:
                publicId:
                  type: string
                  description: Reporter's 9-digit publicId (optional; anomalies before login are recorded as anon)
                platform:
                  type: string
                  enum:
                    - web
                    - wechat
                    - crazygames
                buildVersion:
                  type: string
                  description: Client build version (__NW_BUILD_VERSION__, short commit hash; '0.0.0' if unbaked). Lets a recurring anomaly be attributed to a specific deploy, e.g. to rule out a long-open tab still running pre-fix code.
                device:
                  type: string
                  enum:
                    - phone
                    - tablet
                    - desktop
                    - unknown
                  description: 'Coarse hardware bucket, derived client-side (net/anomaly/deviceContext.ts). Distinct from `platform`, which is the BUILD TARGET — a phone browser and a desktop browser both report platform=web, so before this field existed "it only crashes on mobile" could not be confirmed or counted from Loki at all. Deliberately a 4-value bucket rather than the raw UA: this lands inline on every Loki line, where a free-form string would be both high-cardinality and a privacy liability. Server allowlists the value.'
                dpr:
                  type: number
                  description: Physical pixels per CSS px, rounded to 2dp (server clamps to 0..16)
                mem:
                  type: number
                  description: Approximate device RAM in GB (navigator.deviceMemory; Chromium-only, absent elsewhere). Separates "only ever on low-memory hardware" from "everywhere" for OOM-class reports. Server clamps to 0..1024.
                events:
                  type: array
                  description: A batch of anomaly events; server takes at most the first 200
                  items:
                    type: object
                    required:
                      - type
                      - msg
                      - ts
                    properties:
                      type:
                        type: string
                        description: Anomaly type (mem/cpu/webgl_lost/anr/jserror/crash; unknown falls back to other)
                      msg:
                        type: string
                      ts:
                        type: number
                        description: epoch ms (client clock)
                      detail:
                        type: string
                        description: Structured supplement (already compressed into a single string and truncated on the client)
                      orient:
                        type: string
                        enum:
                          - portrait
                          - landscape
                        description: 'Orientation when this event was reported. Per-event, not per-batch: a `crash` event describes the orientation of the session that DIED (restored from the sentinel), while events queued alongside it describe the live one. Server allowlists the value.'
                      vp:
                        type: string
                        description: Viewport as `WxH` in CSS px at report time (the box the game lays out against — not `screen.*`, which does not shrink for the keyboard or browser chrome)
                      sinceRot:
                        type: number
                        description: ms between the last orientation flip and this event; absent when the session never rotated. The signal that makes rotation-time failures distinguishable rather than merely suspected — such reports cluster at sinceRot≈0 while ordinary ones carry no sinceRot at all. Server clamps to 0..86400000.
      responses:
        '200':
          description: Accepted (accepted = actual forwarded count; 0 if rate-limited or invalid)
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - data
                properties:
                  ok:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: object
                    required:
                      - accepted
                    properties:
                      accepted:
                        type: integer
        '400':
          $ref: '#/components/responses/ErrorResp'
...
  /analytics/events:
    post:
      tags:
        - analytics
      operationId: postAnalyticsEvents
      summary: Batch event reporting (JWT optional; attaches user_id when token present, otherwise anonymous)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsEventBatch'
      responses:
        '200':
          description: Accepted (fire-and-forget, does not guarantee persistence)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum:
                      - true
                  data:
                    nullable: true